--- storage/badger/operation/job_consume.go
@@ -1,17 +1,0 @@
-package operation
-
-import "github.com/dgraph-io/badger/v2"
-
-// RetrieveProcessedIndex returns the processed index for a job consumer
-func RetrieveProcessedIndex(jobName string, processed *int64) func(*badger.Txn) error {
-	return retrieve(makePrefix(codeJobConsumerProcessed, jobName), processed)
-}
-
-func InsertProcessedIndex(jobName string, processed int64) func(*badger.Txn) error {
-	return insert(makePrefix(codeJobConsumerProcessed, jobName), processed)
-}
-
-// SetProcessedIndex updates the processed index for a job consumer with given index
-func SetProcessedIndex(jobName string, processed int64) func(*badger.Txn) error {
-	return update(makePrefix(codeJobConsumerProcessed, jobName), processed)
-}
