+++ storage/badger/operation/jobs.go
@@ -27,3 +27,17 @@ func RetrieveJobAtIndex(queue string, index int64, entity *flow.Identifier) func
 func InsertJobAtIndex(queue string, index int64, entity flow.Identifier) func(*badger.Txn) error {
 	return insert(makePrefix(codeJobQueue, queue, uint64(index)), entity)
 }
+
+// RetrieveProcessedIndex returns the processed index for a job consumer
+func RetrieveProcessedIndex(jobName string, processed *int64) func(*badger.Txn) error {
+	return retrieve(makePrefix(codeJobConsumerProcessed, jobName), processed)
+}
+
+func InsertProcessedIndex(jobName string, processed int64) func(*badger.Txn) error {
+	return insert(makePrefix(codeJobConsumerProcessed, jobName), processed)
+}
+
+// SetProcessedIndex updates the processed index for a job consumer with given index
+func SetProcessedIndex(jobName string, processed int64) func(*badger.Txn) error {
+	return update(makePrefix(codeJobConsumerProcessed, jobName), processed)
+}
